examples/graphql/requirements.txt
strawberry-graphql[fastapi]>=0.219.0
uvicorn[standard]>=0.32.0
surrealdb>=1.0.6
python-dotenv>=1.0.0

